Review fix: move Op related functions from chain.ex
