+++ apps/explorer/lib/explorer/chain/optimism/deposit.ex
@@ -3,7 +3,12 @@ defmodule Explorer.Chain.Optimism.Deposit do
 
   use Explorer.Schema
 
+  import Explorer.Chain, only: [join_association: 3, select_repo: 1]
+
   alias Explorer.Chain.{Hash, Transaction}
+  alias Explorer.PagingOptions
+
+  @default_paging_options %PagingOptions{page_size: 50}
 
   @required_attrs ~w(l1_block_number l1_transaction_hash l1_transaction_origin l2_transaction_hash)a
   @optional_attrs ~w(l1_block_timestamp)a
@@ -47,6 +52,35 @@ defmodule Explorer.Chain.Optimism.Deposit do
       select: {d.l1_block_number, d.l1_transaction_hash},
       order_by: [desc: d.l1_block_number],
       limit: 1
+    )
+  end
+
+  @doc """
+  Lists `t:Explorer.Chain.Optimism.Deposit.t/0`'s' in descending order based on l1_block_number and l2_transaction_hash.
+
+  """
+  @spec list :: [__MODULE__.t()]
+  def list(options \\ []) do
+    paging_options = Keyword.get(options, :paging_options, @default_paging_options)
+
+    base_query =
+      from(d in __MODULE__,
+        order_by: [desc: d.l1_block_number, desc: d.l2_transaction_hash]
+      )
+
+    base_query
+    |> join_association(:l2_transaction, :required)
+    |> page_deposits(paging_options)
+    |> limit(^paging_options.page_size)
+    |> select_repo(options).all()
+  end
+
+  defp page_deposits(query, %PagingOptions{key: nil}), do: query
+
+  defp page_deposits(query, %PagingOptions{key: {block_number, l2_tx_hash}}) do
+    from(d in query,
+      where: d.l1_block_number < ^block_number,
+      or_where: d.l1_block_number == ^block_number and d.l2_transaction_hash < ^l2_tx_hash
     )
   end
 end
